requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
torch==2.0.1
scikit-learn==1.3.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
pymupdf==1.22.5
Werkzeug==2.3.7
